Did some minor adjustments to FGM and DFGM, in particular, added check of norm to prevent quotient from vanishing when the input is really sparse like url!
Running data for all days, some observations:
the scalability of a distributed version should stem from two sources, the first one being number of samples and the second being number of kernels
Performance takes a little bit of a hit when the input feature is of really really high dimension because we need to compute a full weight vector in the most violated research stage.
Luckily we only need to do it evey time a kernel is added and the number of kernel is usually smaller than 10, which is why we can still see significant speedup, especially when the input dimension is not high but number of samples is.
For FGM I set its max_loop to 12, while for DFGM I set it to 4. While this may seem unfair, it seems that DFGM can approximate the results quite good, so may be the second order hessian approximation can be thought of as another form of "stochastic gradient descent" in the sense that it gives us (or utilize) uncertainties.
Note also that DFGM is not necessarily slower than FGM even in a single machine case. I will need to analyze the data to see why in some cases this is true though.